% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_graphs.R
\name{calc_average_causal_effect_estimand}
\alias{calc_average_causal_effect_estimand}
\title{Calculate average causal effect estimand}
\usage{
calc_average_causal_effect_estimand(DATA, conditioning_variable)
}
\arguments{
\item{DATA}{a dataset with potential outcomes named y0 and y1}
}
\value{
number
}
\description{
DETAIL
}

