% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_graphs.R
\name{calc_conditional_difference_in_means}
\alias{calc_conditional_difference_in_means}
\title{Calculate the naive difference in means}
\usage{
calc_conditional_difference_in_means(DATA, conditioning_variable)
}
\arguments{
\item{DATA}{a dataset, can't have a variable named "Z" in it}

\item{conditioning_variable}{character object for variable name in DATA;
assumed to be binary 0-1}
}
\value{
number
}
\description{
Calculate the naive difference in means
}

