% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potential_outcomes.R
\name{calc_naive_difference_in_means}
\alias{calc_naive_difference_in_means}
\title{Calculate the naive difference in means}
\usage{
calc_naive_difference_in_means(DATA)
}
\arguments{
\item{DATA}{a dataset}
}
\value{
number
}
\description{
Calculate the naive difference in means
}

