% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potential-outcomes.R
\name{calculate_naive_difference_in_means}
\alias{calculate_naive_difference_in_means}
\title{Calculates the naive difference in means}
\usage{
calculate_naive_difference_in_means(DATA)
}
\arguments{
\item{DATA}{a dataset}
}
\value{
number
}
\description{
Calculates the naive difference in means
}

