% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_dataset}
\alias{load_dataset}
\title{Loads a dataset from inst/extdata into memory.}
\usage{
load_dataset(dataset)
}
\arguments{
\item{dataset}{the name of the dataset to load}
}
\value{
data set
}
\description{
Loads a dataset from inst/extdata into memory.
}

