% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_graphs.R
\name{make_data_for_figure_4_10}
\alias{make_data_for_figure_4_10}
\title{Simulates some data for a thingy}
\usage{
make_data_for_figure_4_10(n, seed = sample.int(.Machine$integer.max, 1))
}
\arguments{
\item{n}{number of observations}

\item{seed}{integer value for random seed}
}
\value{
data.table with simulated data
}
\description{
DETAILS
William came up with it, don't ask me
}

