% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_observational_data}
\alias{simulate_observational_data}
\title{Simulate an observational study}
\usage{
simulate_observational_data(n, seed = sample.int(.Machine$integer.max, 1))
}
\arguments{
\item{n}{number of individuals}

\item{seed}{integer value for random seed}
}
\value{
data.table with potential outcomes
}
\description{
This study will all individuals to select their treatment value
based on their potential outcomes
}

