% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potential_outcomes.R
\name{simulate_simple_experiment}
\alias{simulate_simple_experiment}
\title{Simulate a simple experiment}
\usage{
simulate_simple_experiment(n, prob_treatment = 0.5,
  seed = sample.int(.Machine$integer.max, 1))
}
\arguments{
\item{n}{number of individuals}

\item{prob_treatment}{probability of treatment}

\item{seed}{integer value for random seed}
}
\value{
data.table with potential outcomes
}
\description{
Simulates potential outcomes and adds a treatment variable
}

