% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{calc_naive_difference_in_means}
\alias{calc_naive_difference_in_means}
\title{Calculate naive difference in means}
\usage{
calc_naive_difference_in_means(DATA)
}
\arguments{
\item{DATA}{data.table generated by simulate_simple_experiment}
}
\value{
estimated average
}
\description{
Calculate naive difference in means
}

