% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_dataset}
\alias{load_dataset}
\title{Load datasets from inst/extdata folder into the memory}
\usage{
load_dataset(dataset)
}
\arguments{
\item{dataset}{the name of the dataset to load}
}
\value{
dataset
}
\description{
Load datasets from inst/extdata folder into the memory
}

