% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{make_data_for_figure_4_1}
\alias{make_data_for_figure_4_1}
\title{Simulate data that corresponds to Figure 4.1 from Morgan & Winship (2007)}
\usage{
make_data_for_figure_4_1(n, seed = sample.int(.Machine$integer.max, 1))
}
\arguments{
\item{n}{number of observations}

\item{seed}{RNG seed}
}
\value{
data.table with potential outcomes
}
\description{
Data generation
}

