% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potential-outcomes.R
\name{simulate_potential_outcomes}
\alias{simulate_potential_outcomes}
\title{Makes binary causal states and binary valued potential outcomes.}
\usage{
simulate_potential_outcomes(n, seed = sample.int(.Machine$integer.max, 1))
}
\arguments{
\item{n}{number of observations}

\item{seed}{RNG seed}
}
\value{
data.table with potential outcomes
}
\description{
Makes binary causal states and binary valued potential outcomes.
}

