% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_potential_outcomes}
\alias{simulate_potential_outcomes}
\title{Simple simulation about potential outcomes and causal estimation}
\usage{
simulate_potential_outcomes(n, seed = sample.int(.Machine$integer.max, 1))
}
\arguments{
\item{n}{number of individuals}

\item{seed}{integer value for random seed}
}
\value{
data.table with potential outcomes
}
\description{
Make a binary causal state, a binary valued potential outcome
}

