% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potential-outcomes.R
\name{simulate_simple_experiment}
\alias{simulate_simple_experiment}
\title{Simulates potential outcomes and adds a treatment variable,
assumes SUTVA, defines the observed outcomes, deletes potential outcomes, and
defines a result.}
\usage{
simulate_simple_experiment(n, prob_treatment = 0.5,
  seed = sample.int(.Machine$integer.max, 1))
}
\arguments{
\item{n}{number of observations}

\item{prob_treatment}{probability of treatment}

\item{seed}{RNG seed}
}
\value{
data.table with experiment data
}
\description{
Simulates potential outcomes and adds a treatment variable,
assumes SUTVA, defines the observed outcomes, deletes potential outcomes, and
defines a result.
}

